requirements.txt
gevent==1.0.2
greenlet==0.4.13
ProxyTypes==0.9
nose==1.3.0
wsgiref==0.1.2
feedparser==5.1.3
Jinja2==2.7
Werkzeug==0.9.3
Flask==0.10.1
Flask-Cors==3.0.2
redis==2.7.6
msgpack-python==0.1.12
python-dateutil==2.1
gunicorn
bugsnag
blinker
python-slugify
